requirements.txt
Django
gunicorn
django-heroku
dj-database-url
django-allauth
psycopg2
djangorestframework
requests
django-crispy-forms